Historia terminada, solo falta hacer Mergue con la rama principal

+++ web/src/main/java/com/portal/formula1/service/PilotoService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -13,11 +14,21 @@ public class PilotoService {
     @Autowired
     private PilotoDAO pilotoDAO;
 
+    @Autowired
+    private EncuestaService encuestaService;
+
     public Piloto guardarPiloto(Piloto piloto) {
         return pilotoDAO.save(piloto);
     }
 
     public void eliminarPiloto(Integer dorsal) {
+        Piloto piloto = pilotoDAO.findById(dorsal)
+                .orElseThrow(() -> new NoSuchElementException("Piloto no encontrado"));
+
+        if (encuestaService.estaPilotoEnEncuestaActiva(piloto.getDorsal())) {
+            throw new IllegalStateException("El piloto está en una encuesta activa y no puede ser eliminado.");
+        }
+
         pilotoDAO.deleteById(dorsal);
     }
